refactor: rename all layer style files to use `LayerStyle` in filenames and when referenced in code
Related pull requests: #102, #103
